requirements.txt
alpha-vantage==2.1.3
astroid==2.3.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
python-dotenv==0.12.0
yfinance
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
xlrd
torchvision
torch
pandas
mplfinance
Pillow
matplotlib
numpy
psycopg2

